module name change and lint fixes
