--- HFStrategy/__init__.py
@@ -1,6 +1,0 @@
-name = 'HFStrategy'
-
-from HFStrategy.Strategy.Strategy import Strategy
-from HFStrategy.Strategy.PositionManager import PositionError
-from HFStrategy.Strategy.Position import Position
-from HFStrategy.utils.Executor import *
