--- HFStrategy/models/__init__.py
@@ -1,3 +1,0 @@
-
-from .events import *
-from .price_update import *
