+++ hfstrategy/models/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module contains useful objects which are used to help type
+different data models.
+"""
+
+from .events import Events
+from .price_update import PriceUpdate
