+++ pylint.rc
@@ -1,1 +1,6 @@
 [MESSAGES CONTROL]
+
+indent-string='  '
+disable=too-few-public-methods,
+        import-error,
+        too-many-arguments
